Live Swapping (#566)
Adds the ability to swap between emulators without closing down StandaloneRTC.exe. This gives us the benefit of not having to save/close any work that is currently ongoing in RTC, and opens up the ability to swap implementations gracefully without using /taskkill. It also allows cross emulator stockpiles to be created, as it can seamlessly swap between emulators and load stockpiles.